pyproject.toml
[project]
name = "data-exploration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen-agentchat>=0.7.4",
    "autogen-ext[anthropic,azure,bedrock,openai]>=0.7.4",
    "azure-identity>=1.25.0",
    "boto3>=1.40.30",
    "datetime>=5.5",
    "dotenv>=0.9.9",
    "htbuilder>=0.9.0",
    "langchain-aws>=0.2.33",
    "langchain-core>=0.3.76",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "logging>=0.4.9.6",
    "panda>=0.3.1",
    "pandas>=2.3.2",
    "pylint>=3.3.8",
    "requests>=2.32.5",
    "snowflake-sqlalchemy>=1.7.7",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.49.1",
    "tabulate>=0.9.0",
]
